+++ sickrage/providers/__init__.py
@@ -64,7 +64,7 @@ class GenericProvider(object):
         self.supports_backlog = False
         self.supports_absolute_numbering = False
         self.anime_only = False
-        self.search_mode = None
+        self.search_mode = 'eponly'
         self.search_fallback = False
         self.enabled = False
         self.enable_daily = False
@@ -486,7 +486,8 @@ class GenericProvider(object):
         if self.enable_cookies and self.cookies:
             cookie_validator = re.compile(r'^(\w+=\w+)(;\w+=\w+)*$')
             if cookie_validator.match(self.cookies):
-                add_dict_to_cookiejar(sickrage.srCore.srWebSession.cookies, dict(x.rsplit('=', 1) for x in self.cookies.split(';')))
+                add_dict_to_cookiejar(sickrage.srCore.srWebSession.cookies,
+                                      dict(x.rsplit('=', 1) for x in self.cookies.split(';')))
                 return True
 
         return False
@@ -703,8 +704,8 @@ class NZBProvider(GenericProvider):
 
     def __init__(self, name, url, private):
         super(NZBProvider, self).__init__(name, url, private)
-        self.api_key = None
-        self.username = None
+        self.api_key = ''
+        self.username = ''
 
     @property
     def isActive(self):
@@ -823,15 +824,12 @@ class TorrentRssProvider(TorrentProvider):
         return title, url
 
     def validateRSS(self):
-
         try:
             if self.cookies:
                 cookie_validator = re.compile(r"^(\w+=\w+)(;\w+=\w+)*$")
                 if not cookie_validator.match(self.cookies):
                     return False, 'Cookie is not correctly formatted: ' + self.cookies
 
-            # pylint: disable=W0212
-            # Access to a protected member of a client class
             data = self.cache._get_rss_data()['entries']
             if not data:
                 return False, 'No items found in the RSS feed ' + self.urls['base_url']
@@ -883,12 +881,31 @@ class TorrentRssProvider(TorrentProvider):
 
     @classmethod
     def getProviders(cls):
-        return cls.getDefaultProviders()
+        providers = cls.getDefaultProviders()
+
+        try:
+            for curProviderStr in sickrage.srCore.srConfig.CUSTOM_PROVIDERS.split('!!!'):
+                if not len(curProviderStr):
+                    continue
+
+                try:
+                    cur_type, curProviderData = curProviderStr.split('|', 1)
+                    if cur_type == "torrentrss":
+                        cur_name, cur_url, cur_cookies, cur_title_tag = curProviderData.split('|')
+                        cur_url = sickrage.srCore.srConfig.clean_url(cur_url)
+
+                        providers += [TorrentRssProvider(cur_name, cur_url, False, cur_cookies, cur_title_tag)]
+                except Exception:
+                    continue
+        except Exception:
+            pass
+
+        return providers
 
     @classmethod
     def getDefaultProviders(cls):
         return [
-            cls('showRSS', 'showrss.info', False, None, 'title', 'eponly', True, True, True, True)
+            cls('showRSS', 'showrss.info', False, '', 'title', 'eponly', True, True, True, True)
         ]
 
 
@@ -899,7 +916,7 @@ class NewznabProvider(NZBProvider):
                  name,
                  url,
                  private,
-                 key=None,
+                 key='',
                  catIDs='5030,5040',
                  search_mode='eponly',
                  search_fallback=False,
@@ -1145,16 +1162,36 @@ class NewznabProvider(NZBProvider):
 
     @classmethod
     def getProviders(cls):
-        return cls.getDefaultProviders()
+        providers = cls.getDefaultProviders()
+
+        try:
+            for curProviderStr in sickrage.srCore.srConfig.CUSTOM_PROVIDERS.split('!!!'):
+                if not len(curProviderStr):
+                    continue
+
+                try:
+                    cur_type, curProviderData = curProviderStr.split('|', 1)
+
+                    if cur_type == "newznab":
+                        cur_name, cur_url, cur_key, cur_cat = curProviderData.split('|')
+                        cur_url = sickrage.srCore.srConfig.clean_url(cur_url)
+
+                        providers += [NewznabProvider(cur_name, cur_url, bool(not cur_key == 0), key=cur_key)]
+                except Exception:
+                    continue
+        except Exception:
+            pass
+
+        return providers
 
     @classmethod
     def getDefaultProviders(cls):
         return [
-            cls('SickBeard', 'lolo.sickbeard.com', False, None, '5030,5040', 'eponly', False, False, False, True),
-            cls('NZB.Cat', 'nzb.cat', True, None, '5030,5040,5010', 'eponly', True, True, True, True),
-            cls('NZBGeek', 'api.nzbgeek.info', True, None, '5030,5040', 'eponly', False, False, False, True),
-            cls('NZBs.org', 'nzbs.org', True, None, '5030,5040', 'eponly', False, False, False, True),
-            cls('Usenet-Crawler', 'usenet-crawler.com', True, None, '5030,5040', 'eponly', False, False, False, True)
+            cls('SickBeard', 'lolo.sickbeard.com', False, '', '5030,5040', 'eponly', False, False, False, True),
+            cls('NZB.Cat', 'nzb.cat', True, '', '5030,5040,5010', 'eponly', True, True, True, True),
+            cls('NZBGeek', 'api.nzbgeek.info', True, '', '5030,5040', 'eponly', False, False, False, True),
+            cls('NZBs.org', 'nzbs.org', True, '', '5030,5040', 'eponly', False, False, False, True),
+            cls('Usenet-Crawler', 'usenet-crawler.com', True, '', '5030,5040', 'eponly', False, False, False, True)
         ]
 
 
@@ -1223,27 +1260,18 @@ class providersDict(dict):
     def __init__(self):
         super(providersDict, self).__init__()
 
-        self.filename = os.path.abspath(os.path.join(sickrage.DATA_DIR, 'providers.db'))
+        self.provider_order = []
 
+        self[NZBProvider.type] = {}
+        self[TorrentProvider.type] = {}
+        self[NewznabProvider.type] = {}
+        self[TorrentRssProvider.type] = {}
+
+    def sync(self):
         self[NZBProvider.type] = dict([(p.id, p) for p in NZBProvider.getProviders()])
         self[TorrentProvider.type] = dict([(p.id, p) for p in TorrentProvider.getProviders()])
         self[NewznabProvider.type] = dict([(p.id, p) for p in NewznabProvider.getProviders()])
         self[TorrentRssProvider.type] = dict([(p.id, p) for p in TorrentRssProvider.getProviders()])
-
-        self.provider_order = []
-        self.sort()
-
-    def sync(self):
-        remove = []
-
-        # find
-        for p in self.provider_order:
-            if p not in self.all():
-                remove.append(p)
-
-        # remove
-        for r in remove:
-            self.provider_order.pop(self.provider_order.index(r))
 
     def sort(self, key=None, randomize=False):
         sorted_providers = []
@@ -1267,33 +1295,22 @@ class providersDict(dict):
         return dict([(pID, pObj) for pID, pObj in self.all().items() if not pObj.isEnabled])
 
     def all(self):
-        return reduce(lambda a, b: a.update(b) or a, [
-            self.nzb(),
-            self.torrent(),
-            self.newznab(),
-            self.torrentrss()
-        ], {})
+        return dict(self.nzb().items() + self.torrent().items() + self.newznab().items() + self.torrentrss().items())
 
     def all_nzb(self):
-        return reduce(lambda a, b: a.update(b) or a, [
-            self.nzb(),
-            self.newznab()
-        ], {})
+        return dict(self.nzb().items() + self.newznab().items())
 
     def all_torrent(self):
-        return reduce(lambda a, b: a.update(b) or a, [
-            self.torrent(),
-            self.torrentrss()
-        ], {})
+        return dict(self.torrent().items() + self.torrentrss().items())
 
     def nzb(self):
         return self[NZBProvider.type]
 
-    def newznab(self):
-        return self[NewznabProvider.type]
-
     def torrent(self):
         return self[TorrentProvider.type]
+
+    def newznab(self):
+        return self[NewznabProvider.type]
 
     def torrentrss(self):
         return self[TorrentRssProvider.type]